Wandile/feature/api model classes

+++ sql/loan.sql
@@ -8,7 +8,7 @@ CREATE TYPE loan_type_enum AS ENUM ('short-term', 'long-term');
 CREATE TABLE loan (
     loan_id UUID DEFAULT gen_random_uuid() PRIMARY KEY,
     account_id UUID NOT NULL,
-    loan_amount NUMERIC(13, 3) NOT NULL,
+    loan_amount BIGINT NOT NULL,
     loan_type loan_type_enum NOT NULL
 );
 -- rollback DROP TABLE loan
